requirements.txt
asyncpg==0.25.0
black==22.1.0
isort==5.10.1
PyJWT==2.3.0
python-dotenv==0.20.0
quart==0.17.0
uvloop==0.16.0
